Now using genreflex. Finalised TrackPar interface
