--- interface/LinkDef.h
@@ -1,5 +1,0 @@
-#ifdef __CLING__
-#pragma link C++ class InterpolatorFunc+;
-#pragma link C++ class WFViewer+;
-#pragma link C++ class MCPAnalyzer+;
-#endif
